Database.SQLFragment.Proxies

Description

Provides empty typed SQLFragments to force fragment type signature using !::! and !#:! operators. This module is meant to be import as qualified as some names can clashes with standard names.

>>> :set -XOverloadedStrings
>>> :t "a" !::! string
"a" !::! string :: SQLFragment '[String] '[]

Documentation

empty :: SQLFragment [] []

string :: SQLFragment ((:) * String ([] *)) ([] *)

int :: SQLFragment ((:) * Int ([] *)) ([] *)

double :: SQLFragment ((:) * Double ([] *)) ([] *)