Database.SQLFragment.SQLFragment

Types

data Section

Constructors

COLUMN

FROM

WHERE

GROUP

HAVING

ORDER

data SQLFragment e p

Represents a bit of an SQL Query. Could be equivalent to

"name FROM user"
"FROM user WHERE name = Smith
"GROUP BY name ORDER BY name"

Fragments can be combined to form a full query. However, the type of the query "SELECT", "INSERT" is not carried by the fragment but decided when generating the query. Fragments also carry some phantom types which are used to type the query results and its parameters if any. Phantom types allow type safe query composition even though SQLFragment is designed to not be "on the way", and so can be retyped easily.

Fragments can be left untyped by simply using "SQLFragment '[] '[]"

SQLFragment can be constructing from a string using the OverloadedStrings extension.

Accepted format are :

column

table.column

table [AS] alias.column

table.

table [AS] alis.

database.table.column

>>> :set -XOverloadedStrings

>>> toSelectQuery "a"
"SELECT a"
>>> toSelectQuery "t.a"
"SELECT t.a\nFROM t"
>>> toSelectQuery "t as t2."
"FROM t as t2"
>>> toSelectQuery "db.t as t2.a"
"SELECT t2.a\nFROM db.t as t2"

Constructors

SQLFragment

Fields

clauses :: Map Section [String]

Query Generations

toSelectQuery :: SQLFragment e p -> String

toInsertQuery :: SQLFragment e p -> String

Insert using VALUES and ?

>>> toInsertQuery' $ "t.a"
"INSERT INTO t (t.a) VALUES (?)"

Types Manipulation

The following functions change the type signature of a fragment. Combined with OverloadedStrings, they can be used to create a fragment from a string.

forgetTypes :: SQLFragment e p -> SQLFragment e' p'

Forget the phantom types to whatever is needed.

>>> :set -XDataKinds
>>> let a = "a" :: SQLFragment '[] '[]
>>> let b = forgetTypes a :: SQLFragment '[String] '[]

clearTypes :: SQLFragment e p -> SQLFragment [] []

nullable :: SQLFragment e p -> SQLFragment (Nullable e) (Nullable p)

Change types to Maybe version if needed

notnull :: SQLFragment e p -> SQLFragment (NotNull e) (NotNull p)

Joins

data JoinType

Type of Join. At the moment only left and inner join are supported.

Constructors

LEFT

RIGHT

INNER

OUTER

data Join

Join between 2 tables.

Constructors

Join

Fields

source :: String

destination :: String

joinClause :: Maybe String

joinType :: JoinType

innerJoin :: Boolables a => SQLFragment a ([] *) -> Join

leftJoin :: Boolables a => SQLFragment a ([] *) -> Join

rightJoin :: Boolables a => SQLFragment a ([] *) -> Join

outerJoin :: Boolables a => SQLFragment a ([] *) -> Join

open :: Join -> Join

openLeft :: Join -> Join

openRight :: Join -> Join

close :: Join -> Join

closeLeft :: Join -> Join

closeRight :: Join -> Join

reverseJoinType :: Join -> Join

inverseJoinType :: Join -> Join

Miscellaneous

bracketize :: String -> String

qualify :: String -> String -> String

>>> qualify "table" "column"
"table.column"

aliasWithPrefix

Arguments

:: String
Alias suffix

-> String
aliased

-> String

Alias a string with a prefix

>>> aliasWithPrefix "2" "table"
"table table2"

from :: String -> SQLFragment [] []